% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapredict_confusion.R
\name{calcConfusionCv}
\alias{calcConfusionCv}
\title{Calculate confusion matrix for cross-validation.}
\usage{
calcConfusionCv(
  cvFit,
  lambda,
  ematMerged,
  sampleMetadata,
  className = "class",
  classLevels = NA
)
}
\arguments{
\item{cvFit}{cv.glmnet object from \code{\link[=metapredictCv]{metapredictCv()}}.}

\item{lambda}{value of lambda at which to use predictions.}

\item{ematMerged}{matrix of gene expression for genes by samples.}

\item{sampleMetadata}{data.frame of sample metadata.}

\item{className}{name of column in \code{sampleMetadata} containing the true
labels.}

\item{classLevels}{Order of classes in the confusion matrix. If \code{NA}
(default), then the function uses the order in \code{cvFit}.}
}
\value{
An object of class \code{table}.
}
\description{
Calculate a confusion matrix based on predictions from cross-validation.
}
