% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapredict_process.R
\name{downloadCustomCdfMappings}
\alias{downloadCustomCdfMappings}
\title{Download custom CDF mapping files from Brainarray.}
\usage{
downloadCustomCdfMappings(cdf, path = ".", ver = 23)
}
\arguments{
\item{cdf}{\verb{data frame} with columns \code{download} (e.g., 'Mouse4302_Mm_ENTREZ')
and \code{rename} (e.g., 'mouse4302mmentrezgcdf').}

\item{ver}{integer version number (23 as of 23 Nov 2018).}
}
\description{
Download Brainarray custom CDF mapping files, which are used for mapping probes to genes
in datasets whose \code{studyDataType} is 'affy_series_matrix'. See
\url{http://brainarray.mbni.med.umich.edu/Brainarray/Database/CustomCDF/CDF_download.asp}.
}
