% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluate_Performance_Lasso.R
\name{evaluate_lasso}
\alias{evaluate_lasso}
\title{This function is to take in data, see what modules are present - then randomly sample the data to create incompleteness simulations and predict for all simulations, and make confusion matrices for each one. and store all results in a list.}
\usage{
evaluate_lasso(.data, moduleVector = NULL)
}
\arguments{
\item{.data}{Tibble object created with the read_data function - KEGG orthology annotation data for one or more bacterial genome annotation files.
#' @param moduleVector Character vector. An optional character vector of specific KEGG Modules to scan user annotations for, and optionally to predict for. Default is NULL.}
}
\description{
This function is to take in data, see what modules are present - then randomly sample the data to create incompleteness simulations and predict for all simulations, and make confusion matrices for each one. and store all results in a list.
}
