% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapredict_process.R
\name{getStudyDataList}
\alias{getStudyDataList}
\title{Get the gene expression data for multiple studies.}
\usage{
getStudyDataList(parentFolderPath, studyMetadata)
}
\arguments{
\item{parentFolderPath}{Path to the folder that contains the data.}

\item{studyMetadata}{A data.frame where each row corresponds to one study.
rownames should correspond to study names, with columns for
studyDataType and platformInfo.}
}
\value{
A named list of ExpressionSets.
}
\description{
\code{getStudyDataList} runs \code{getStudyData} for multiple studies.
}
