% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapredict_process.R
\name{getUnsupportedPlatforms}
\alias{getUnsupportedPlatforms}
\title{Get the GPLs for unsupported microarray platforms.}
\usage{
getUnsupportedPlatforms(studyMetadata)
}
\arguments{
\item{studyMetadata}{\code{data.frame} of study metadata, with columns for
\code{studyDataType} and \code{platformInfo}.}
}
\value{
A vector of platforms for which \code{studyDataType} is 'series_matrix'
and \code{platformInfo} is not supported.
}
\description{
Get the GPLs for unsupported microarray platforms.
}
