% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapredict_predict.R
\name{makeCoefDt}
\alias{makeCoefDt}
\title{Make data.table of non-zero coefficients from a glmnet model.}
\usage{
makeCoefDt(fitResult, lambda, decreasing = TRUE, classLevels = NA)
}
\arguments{
\item{fitResult}{\code{glmnet} object.}

\item{lambda}{value of lambda for which to obtain coefficients.}

\item{decreasing}{logical passed to \code{order}.}

\item{classLevels}{order of columns in resulting \code{data.table}.}
}
\value{
A \code{data.table}.
}
\description{
Make a sorted \code{data.table} of the non-zero coefficients of a logistic or
multinomial \code{glmnet} model.
}
