% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapredict_merge.R
\name{mergeStudyData}
\alias{mergeStudyData}
\title{Merge gene expression data from multiple studies.}
\usage{
mergeStudyData(
  ematList,
  sampleMetadata,
  batchColname = "study",
  covariateName = NA,
  batchCorrection = TRUE,
  parPrior = TRUE
)
}
\arguments{
\item{ematList}{list of expression matrices.}

\item{sampleMetadata}{data.frame for sample metadata.}

\item{batchColname}{column in sampleMetadata containing
batch information for ComBat.}

\item{covariateName}{column in sampleMetadata containing
additional covariates for ComBat besides batch.}

\item{batchCorrection}{\code{TRUE} indicates cross-study normalization
will be performed, \code{FALSE} indicates datasets will be merged
without performing normalization.}

\item{parPrior}{passed to ComBat's \code{par.prior} argument.}
}
\value{
A matrix of expression for genes by samples.
}
\description{
Merge gene expression data from multiple studies, keeping only genes measured
in each dataset, then optionally perform cross-study normalization using
\code{\link[sva:ComBat]{sva::ComBat()}}.
}
