% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_methods.R
\name{metapredict,MetaPredictInput-method}
\alias{metapredict,MetaPredictInput-method}
\title{This function is used to reconstruct metabolic pathways based on present KEGG Orthology terms. For any
incomplete pathways with missing reactions, it will then calculate the probability that each missing reaction
is present in the input genome but was missed in the sampling process. It takes in user input in the form of
an object created with a read_data function call.}
\usage{
\S4method{metapredict}{MetaPredictInput}(
  object,
  module_vector = names(all_models),
  module_detect_type = c("extract", "detect"),
  output_dir = NULL,
  output_prefix = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{object}{Tibble object created with the read_data function - KEGG orthology annotation data for one or more bacterial genome annotation files.}

\item{module_vector}{Character vector. An optional character vector of specific KEGG Modules to scan user annotations for, and optionally to predict for. Default is a character vector containing identifiers for all available modules.}

\item{module_detect_type}{Character. Method 'extract' will give detailed output component containing all extracted KO gene identifiers associated with KEGG modules. Method 'detect' will only detect presence/absence of modules and will not provide detailed output in results.}

\item{output_dir}{Character. The full or relative path to an output directory where result and summary output will be saved as TSV flatfiles}

\item{output_prefix}{Character. Optional string to prefix to MetaPredict output files. Default is NULL, resulting in output files with default names.}

\item{overwrite}{Logical. If TRUE, output files from running the MetaPredict function that have the same name as existing files in the output directory will overwrite those existing files. Default is FALSE.}
}
\description{
This function is used to reconstruct metabolic pathways based on present KEGG Orthology terms. For any
incomplete pathways with missing reactions, it will then calculate the probability that each missing reaction
is present in the input genome but was missed in the sampling process. It takes in user input in the form of
an object created with a read_data function call.
}
