% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_MetaPredict.R
\name{new_MetaPredict}
\alias{new_MetaPredict}
\title{This function is used to reconstruct metabolic pathways based on present KEGG Orthology terms. For any
incomplete pathways with missing reactions, it will then calculate the probability that each missing reaction
is present in the input genome but was missed in the sampling process. It takes in user input in the form of
an object created with a read_data function call.}
\usage{
new_MetaPredict(userData, taxonList = NULL, cores = 1, output_dir = NULL)
}
\arguments{
\item{userData}{Created with the read_data function - KEGG Orthology data for one or more bacteria/archaea}

\item{taxonList}{A CSV flatfile with lowest taxonomic level of each input organism. Same order as in userData, one per line}

\item{cores}{Number of cores to use for computation. Default is 1}

\item{output_dir}{The full or relative path to an output directory where result and summary output will be saved as TSV flatfiles}
}
\description{
This function is used to reconstruct metabolic pathways based on present KEGG Orthology terms. For any
incomplete pathways with missing reactions, it will then calculate the probability that each missing reaction
is present in the input genome but was missed in the sampling process. It takes in user input in the form of
an object created with a read_data function call.
}
