% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_read_genome_data.R
\name{new_read_genome_data}
\alias{new_read_genome_data}
\title{This function is used to read in user data in the form of a one or more flatfiles, default is TSV format.
Each file should end in "-ko.tsv" and contain a column of KEGG Orthology terms with column title 'KO'
and a column with their associated HMM or Blast E-values titled 'E-value'.}
\usage{
new_read_genome_data(
  Data = NULL,
  File = NULL,
  filePath = NULL,
  filePattern = "-ko.tsv",
  kofamscan = TRUE,
  evalue = 0.001,
  delim = "\\t"
)
}
\arguments{
\item{Data}{A dataframe or list of dataframes of HMM/Blast results which include KEGG Orthology terms and E-values in two of the columns of each dataframe}

\item{File}{The full or relative path to a single input flatfile}

\item{filePath}{The full path to the input file or files}

\item{filePattern}{The file extension pattern. If using multiple files, this will read in all files in the listed directory}

\item{kofamscan}{If the input file or files are output from Kofamscan, set this argument to TRUE, otherwise FALSE. Default is TRUE}

\item{evalue}{The desired E-value cutoff. Default is 0.001}

\item{delim}{The delimiter of the input files. Default is tab}
}
\description{
This function is used to read in user data in the form of a one or more flatfiles, default is TSV format.
Each file should end in "-ko.tsv" and contain a column of KEGG Orthology terms with column title 'KO'
and a column with their associated HMM or Blast E-values titled 'E-value'.
}
