% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_read_metagenome_data.R
\name{new_read_metagenome_data}
\alias{new_read_metagenome_data}
\title{This function is used to read in user metagenome data gene prediction data with associated taxonomic annotations,
and KEGG Orthology HMM/Blast data. Both inputs must be in the form of flatfiles, default is TSV format.}
\usage{
new_read_metagenome_data(
  gene_input,
  ko_input,
  evalue = 0.001,
  gene_delim = "\\t",
  ko_delim = "\\t",
  kofamscan = TRUE
)
}
\arguments{
\item{gene_input}{The full or relative path to a single flatfile with gene predictions and their associated taxonomic annotations.
Columns with taxonomy should include one more of the following: Genus, Family, Order, Class, Phylum, Domain}

\item{ko_input}{The full or relative path to a  single flatfile with HMM/Blast results which include KEGG Orthology terms in a column titled 'KO'
and E-values in a column titled 'E-value'}

\item{evalue}{The desired E-value cutoff for HMM/Blast KEGG Orthology hits. Default is 0.001}

\item{gene_delim}{The delimiter of the input gene prediction with taxonomic annotation flatfile. Default is tab}

\item{ko_delim}{The delimiter of the input KEGG Orthology prediction flatfile. Default is tab}

\item{kofamscan}{If the input KEGG Orthology flatfile is direct output from Kofamscan, set this argument to TRUE, otherwise FALSE. Default is TRUE}
}
\description{
This function is used to read in user metagenome data gene prediction data with associated taxonomic annotations,
and KEGG Orthology HMM/Blast data. Both inputs must be in the form of flatfiles, default is TSV format.
}
