% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapredict_plot.R
\name{plotCoefficients}
\alias{plotCoefficients}
\title{Plot coefficients for a gene expression meta-analysis.}
\usage{
plotCoefficients(fitResult, lambda, classLevels = NA, decreasing = FALSE,
  geneIdOrder = NA, org = "org.Hs.eg")
}
\arguments{
\item{fitResult}{glmnet object.}

\item{lambda}{value of lambda to use in \code{fitResult}.}

\item{classLevels}{Order of classes in the confusion matrix. If \code{NA} (default),
then the function uses the order in \code{fitResult}, passed to \code{makeCoefDf}.}

\item{decreasing}{logical indicating whether to sort genes by decreasing coefficient,
passed to \code{makeCoefDf}.}

\item{geneIdOrder}{Optional character array of Entrez Gene IDs specifying the order
of genes. If \code{NA} (default), the order from \code{makeCoefDf} is used.}

\item{org}{Name of package for mapping Entrez Gene IDs to gene symbols,
passed to \code{data} argument of \code{annotate::lookUp}.}
}
\value{
A ggplot object.
}
\description{
\code{plotCoefficients} generates a ggplot of the coefficients
from a glmnet object trained on a gene expression meta-analysis
(either logistic or multinomial classification).
}
