% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapredict_process.R
\name{downloadCustomCdfMappings}
\alias{downloadCustomCdfMappings}
\title{Download custom CDF mapping files from Brainarray.}
\usage{
downloadCustomCdfMappings(cdf, path = ".", ver = 22)
}
\arguments{
\item{cdf}{data frame with columns \code{download} (e.g., 'Mouse4302_Mm_ENTREZ') and
\code{rename} (e.g., 'mouse4302mmentrezgcdf')}

\item{ver}{integer version number (22 as of 21 Jun 2017)}
}
\description{
Download Brainarray custom CDF mapping files, which are used for mapping probes to genes
in datasets whose \code{studyDataType} is 'affy_series_matrix'. See
\url{http://brainarray.mbni.med.umich.edu/Brainarray/Database/CustomCDF/CDF_download.asp}.
}
