% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapredict_process.R
\name{getStudyData}
\alias{getStudyData}
\title{Get the gene expression data for one study.}
\usage{
getStudyData(parentFolderPath, studyName, studyDataType, platformInfo)
}
\arguments{
\item{parentFolderPath}{Path to the folder that contains the data.}

\item{studyName}{Name of dataset to load.}

\item{studyDataType}{Type of data.}

\item{platformInfo}{Microarray platform.}
}
\value{
An ExpressionSet, unless the platform or data type is not supported, then NA.
}
\description{
\code{getStudyData} loads the gene expression data for one study,
	maps probes to Entrez Gene IDs, and normalizes and transforms
	the expression values.
}
