% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapredict_process.R
\name{getStudyDataList}
\alias{getStudyDataList}
\title{Get the gene expression data for multiple studies.}
\usage{
getStudyDataList(parentFolderPath, studyMetadata)
}
\arguments{
\item{parentFolderPath}{Path to the folder that contains the data.}

\item{studyMetadata}{A \code{data.frame} where each row corresponds to one study.
Should have columns for \code{study}, \code{studyDataType}, and \code{platformInfo}.}
}
\value{
A named list of \code{ExpressionSet}s.
}
\description{
Runs \code{\link[=getStudyData]{getStudyData()}} for multiple studies.
}
