% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapredict_process.R
\name{getSupportedPlatforms}
\alias{getSupportedPlatforms}
\title{Get the GPLs for microarray platforms that are currently supported.}
\usage{
getSupportedPlatforms()
}
\value{
A character array of currently supported GPLs.
}
\description{
Get the GPLs for microarray platforms that are currently supported.
}

