% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapredict_process.R
\name{getUnsupportedPlatforms}
\alias{getUnsupportedPlatforms}
\title{Get the GPLs for unsupported microarray platforms.}
\usage{
getUnsupportedPlatforms(studyMetadata)
}
\arguments{
\item{studyMetadata}{data.frame of study metadata, with columns for \code{studyDataType} and \code{platformInfo}}
}
\description{
\code{getUnsupportedPlatforms} checks if any studies in studyMetadata
whose \code{studyDataType=='series_matrix'} have \code{platformInfo}
that is not supported.
}

