% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapredict_process.R
\name{installCustomCdfPackages}
\alias{installCustomCdfPackages}
\title{Install custom CDF packages from Brainarray.}
\usage{
installCustomCdfPackages(pkgs, ver = 23)
}
\arguments{
\item{pkgs}{character vector of package names, e.g., 'hgu133ahsentrezgcdf'}

\item{ver}{integer version number (23 as of 23 Nov 2018)}
}
\description{
Install Brainarray custom CDFs for processing raw Affymetrix data. See
\url{http://brainarray.mbni.med.umich.edu/Brainarray/Database/CustomCDF/CDF_download.asp}.
}
