% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapredict_predict.R
\name{makeCoefDf}
\alias{makeCoefDf}
\title{Make data.frame of non-zero coefficients from a glmnet model.}
\usage{
makeCoefDf(fitResult, lambda, decreasing = TRUE, classLevels = NA)
}
\arguments{
\item{fitResult}{glmnet object.}

\item{lambda}{value of lambda for which to obtain coefficients.}

\item{decreasing}{logical passed to \code{order}.}

\item{classLevels}{order of columns in resulting data.frame.}
}
\value{
A data.frame.
}
\description{
\code{makeCoefDf} makes a sorted data.frame of the non-zero
	coefficients of a logistic or multinomial glmnet model.
}
