% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapredict_merge.R
\name{mergeStudyData}
\alias{mergeStudyData}
\title{Merge gene expression data from multiple studies.}
\usage{
mergeStudyData(ematList, sampleMetadata, batchColname = "study",
  covariateName = NA, batchCorrection = TRUE, parPrior = TRUE)
}
\arguments{
\item{ematList}{list of expression matrices.}

\item{sampleMetadata}{data.frame for sample metadata, with rownames
corresponding to sample names.}

\item{batchColname}{column in sampleMetadata containing
batch information for ComBat.}

\item{covariateName}{column in sampleMetadata containing
additional covariates for ComBat besides batch.}

\item{batchCorrection}{TRUE indicates cross-study normalization
will be performed, FALSE indicates datasets will be merged
without performing normalization.}

\item{parPrior}{passed to ComBat's \code{par.prior} argument.}
}
\value{
A matrix of expression for genes by samples.
}
\description{
\code{mergeStudyData} merges gene expression data from
	multiple studies, keeping only genes measured in each
	dataset, then optionally performs cross-study normalization
	using ComBat.
}
