% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lasso_MetaPredict.R
\name{MetaPredict}
\alias{MetaPredict}
\title{This function is used to reconstruct metabolic pathways based on present KEGG Orthology terms. For any
incomplete pathways with missing reactions, it will then calculate the probability that each missing reaction
is present in the input genome but was missed in the sampling process. It takes in user input in the form of
an object created with a read_data function call.}
\usage{
MetaPredict(
  .data,
  moduleVector = NULL,
  predict_models = TRUE,
  predict_type = c("response", "class"),
  output_dir = NULL,
  output_prefix = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{.data}{Created with the read_data function - KEGG Orthology data for one or more bacteria/archaea}

\item{moduleVector}{An optional vector of specific KEGG Modules to scan user annotations and calculate probabilities for}

\item{output_dir}{The full or relative path to an output directory where result and summary output will be saved as TSV flatfiles}
}
\description{
This function is used to reconstruct metabolic pathways based on present KEGG Orthology terms. For any
incomplete pathways with missing reactions, it will then calculate the probability that each missing reaction
is present in the input genome but was missed in the sampling process. It takes in user input in the form of
an object created with a read_data function call.
}
