% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapredict_predict.R
\name{metapredict}
\alias{metapredict}
\title{Predict the response variable in validation datasets.}
\usage{
metapredict(ematList, studyMetadata, sampleMetadata, discoveryStudyNames, alpha,
  lambda, weights, batchColname = "study", covariateName = NA,
  className = "class", type = "response", ...)
}
\arguments{
\item{ematList}{Named list of expression matrices.}

\item{studyMetadata}{data.frame of study metadata.}

\item{sampleMetadata}{data.frame of sample metadata,
with rownames corresponding to sample names.}

\item{discoveryStudyNames}{vector of study names for training.}

\item{alpha}{value of alpha for the elastic net mixing parameter.}

\item{lambda}{value of regularization parameter.}

\item{weights}{vector of weights for training the glmnet model.}

\item{batchColname}{column in sampleMetadata containing
batch information for ComBat.}

\item{covariateName}{column in sampleMetadata containing
additional covariates for ComBat besides batch.}

\item{className}{column in sampleMetadata containing values of the response variable.}

\item{type}{type of prediction to make (passed to \code{predict.glmnet}).}

\item{...}{Other arguments passed to \code{glmnet}.}
}
\value{
A named list of objects from \code{predict.glmnet}.
}
\description{
\code{metapredict} independently merges the discovery datasets
	with each validation dataset, trains a glmnet model on the samples
	from the discovery datasets, then predicts the response variable
	for the samples in the respective validation dataset. This function
	is a wrapper around glmnet.
}
