% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapredict_plot.R
\name{plotClassProbsCv}
\alias{plotClassProbsCv}
\title{Plot class probabilities for samples from cross-validation.}
\usage{
plotClassProbsCv(cvFit, lambda, ematMerged, sampleMetadata,
  className = "class", classLevels = NA, size = 1.5, ggplotArgs = NA)
}
\arguments{
\item{cvFit}{cv.glmnet object from \code{metapredictCv}.}

\item{lambda}{value of lambda to use in \code{fitResult}.}

\item{ematMerged}{matrix of gene expression for genes by samples.}

\item{sampleMetadata}{data.frame of sample metadata.}

\item{className}{column in sampleMetadata containing values of the response variable.}

\item{classLevels}{Preferred order of classes.}

\item{size}{Point size, passed to \code{geom_point}.}

\item{ggplotArgs}{List of additional arguments to add to each ggplot.}
}
\value{
A gtable from \code{gridExtra::arrangeGrob}.
}
\description{
\code{plotClassProbsCv} generates a gtable consisting of multiple
ggplots, one for each dataset used in cross-validation. Within each
class, samples are sorted by the probability of the true class.
}

