% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_genome_data.R
\name{read_genome_data}
\alias{read_genome_data}
\title{This function is used to read in user data in the form of a one or more flatfiles, default is TSV format.
Each file should end in "-ko.tsv" and contain a column of KEGG Orthology terms with column title 'KO'
and a column with their associated HMM or Blast E-values titled 'E-value'.}
\usage{
read_genome_data(
  Data = NULL,
  metadata = NULL,
  metaColnames = TRUE,
  metadata_df = FALSE,
  kofamscan = TRUE,
  dram = FALSE,
  cutoff = 0.001,
  delim = "\\t"
)
}
\arguments{
\item{Data}{A dataframe or list of dataframes of HMM/Blast results which include KEGG Orthology terms and E-values in two of the columns of each dataframe, or the full or relative path to a single or multiple input flatfiles containing this information}

\item{metadata}{A flatfile with one column "taxonomy" that lists the taxonomy of each input genome, and a second column "completeness" that lists the estimated genome completeness of each input genome. The completeness column should be in the format of percent completeness - e.g., a number between 0 and 100 for each genome. Taxonomy should be the NCBI taxonomic name for each genome, from Domain to Species level - it can optionally be "Unknown" if the taxonomic name is not known. The genome_name column should contain the name the user would like to use for each genome. The filename column must contain the full or relative path to each genome annotation file; Note: files can be from different directories.  If metaColnames is TRUE, metadata should have four columns including genome_name, taxonomy, completeness, and filename for each input genome; information for each genome should occupy a single row of the flatfile. Columns can be in any order, and the rows can also be in any order. NOTE: if metaColnames is FALSE - which is not recommended - three unnamed columns will be required that contain the taxonomy, completeness, and filename for each genome.}

\item{metaColnames}{Logcial. If TRUE, the metadata flatfile contains the following columns in any order: genome_name, taxonomy, completeness, filename. If FALSE, the flatfile contains THREE unnamed columns in any order that contain the taxonomy, estimated completeness, and filename for each genome; it is not recommended to use this setting if possible.}

\item{metadata_df}{Metadata can optionally be loaded in as a pre-existing dataframe that contains four columns: genome_name, taxonomy, completeness, and filename. Default is FALSE. Can optionally set metadata_df equal to the object name of a pre-existing metadata dataframe.}

\item{kofamscan}{If the input file or files are output from Kofamscan, set this argument to TRUE, otherwise FALSE. Default is TRUE.}

\item{delim}{The delimiter of the input files. Default is tab.}

\item{evalue}{The desired E-value cutoff. Default is 0.001.}
}
\description{
This function is used to read in user data in the form of a one or more flatfiles, default is TSV format.
Each file should end in "-ko.tsv" and contain a column of KEGG Orthology terms with column title 'KO'
and a column with their associated HMM or Blast E-values titled 'E-value'.
}
